Add bin/graviton for downstream use
The new bin/graviton can be used downstream as vendoro/bin/graviton. It is meant to replace silly vendor/graviton/graviton/app/console calls with something more sane.

For now it is a opy of app/console. If we only ant to expose app/console partially we might change this in the future.